smtp_scan/requirements.txt
Flask==3.2.2
dnspython==3.5.2
python-nmap==0.7.9
